map: redesign API, refactor
Specific map types such as Hashmap and Array don't really work well.
Callers probably have to write wrapper types around them anyway.
Remove these types and, instead, use a single Map type, which
encapsulates all of the possible maps. This way, we cannot enforce
a few additional invariants, but callers have to do that anyway.
So let's be explicit about it and offer the real API.

The Sysfd kludge remains for now, as the rest of the API is figured
out. Some bits (e.g. BTF) are missing, because I know nothing about
them for now.

Signed-off-by: Andrei Tudor Călin <mail@acln.ro>