map, prog: be more helpful when used uninitialized
When clients (in this case, one of the tests) use a Map or Prog without
initializing / loading it first, a panic deep inside package ebpf is
not particularly illuminating.

For a map or a prog to be initialized is equivalent for their attached
mapFD or progFD to be non-nil. Check for that.

Signed-off-by: Andrei Tudor Călin <mail@acln.ro>